+++ gfx/drivers_shader/shader_vulkan.cpp
@@ -1442,7 +1442,7 @@ bool vulkan_filter_chain::init_alias()
    common.texture_semantic_map.clear();
    common.texture_semantic_uniform_map.clear();
 
-   for (i = 0; i < passes.size(); i++)
+   for (i = 0; i < (int)passes.size(); i++)
    {
       unsigned j;
       const std::string name = passes[i]->get_name();
@@ -1472,7 +1472,7 @@ bool vulkan_filter_chain::init_alias()
          return false;
    }
 
-   for (i = 0; i < common.luts.size(); i++)
+   for (i = 0; i < (int)common.luts.size(); i++)
    {
       unsigned j = (unsigned)(&common.luts[i] - common.luts.data());
       if (!slang_set_unique_map(
